pipeline

+++ main.sh
@@ -7,12 +7,15 @@ LANGS=$(ls /home/luca/align-gosact/gosact_chapters_sent/)
 
 for LANG in $LANGS
 do
-  # download language vectors
-  lang_code=$(bash lang_code2lang_model.sh $LANG)
-  bash /home/luca/align-gosact/download_language_model.sh "$lang_code"
-  # align language
-  bash /home/luca/align-gosact/align_lang.sh $LANG
-  # delete language vectors
-  rm /home/luca/word-vectors/cc."$lang_code".300.vec
-  rm /home/luca/word-vectors/cc."$lang_code".300.bin
+  if [ $LANG != "eng" ]
+  then
+    # download language vectors
+    lang_code=$(bash lang_code2lang_model.sh $LANG)
+    bash /home/luca/align-gosact/download_language_model.sh "$lang_code"
+    # align language
+    bash /home/luca/align-gosact/align_lang.sh $LANG
+    # delete language vectors
+    rm /home/luca/word-vectors/cc."$lang_code".300.vec
+    rm /home/luca/word-vectors/cc."$lang_code".300.bin
+  fi
 done
